+++ examples/time_frequency/plot_tfr_topography.py
@@ -41,7 +41,7 @@ epochs = mne.Epochs(raw, events, event_id, tmin, tmax, picks=picks,
                     baseline=(None, 0), reject=dict(grad=4000e-13, eog=150e-6))
 data = epochs.get_data()  # as 3D matrix
 
-layout = mne.layouts.read_layout('Vectorview-all')
+layout = mne.find_layout(epochs.info, 'meg')
 
 ###############################################################################
 # Calculate power and phase locking value