+++ mne/layouts/tests/test_layout.py
@@ -1,10 +1,22 @@
-import os.path as op
-import numpy as np
-from nose.tools import assert_true
-from numpy.testing import assert_array_almost_equal, assert_array_equal
+# Authors: Alexandre Gramfort <gramfort@nmr.mgh.harvard.edu>
+#          Denis Engemann <d.engemann@fz-juelich.de>
+#          Martin Luessi <mluessi@nmr.mgh.harvard.edu>
+#          Eric Larson <larson.eric.d@gmail.com>
+#
+# License: Simplified BSD
 
-from mne.layouts import make_eeg_layout, make_grid_layout, read_layout
-from mne.fiff import Raw
+import copy
+import os.path as op
+import warnings
+
+import numpy as np
+from numpy.testing import assert_array_almost_equal, assert_array_equal
+from nose.tools import assert_true, assert_raises
+
+from mne.layouts import (make_eeg_layout, make_grid_layout, read_layout, 
+                         find_layout)
+
+from mne.fiff import Raw, pick_types, pick_info
 from mne.utils import _TempDir
 
 fif_fname = op.join(op.dirname(__file__), '..', '..', 'fiff',
@@ -12,6 +24,13 @@ fif_fname = op.join(op.dirname(__file__), '..', '..', 'fiff',
 
 lout_path = op.join(op.dirname(__file__), '..', '..', 'fiff',
                     'tests', 'data')
+
+bti_dir = op.join(op.dirname(__file__), '..', '..', 'fiff', 'bti',
+                  'tests', 'data')
+
+fname_ctf_raw = op.join(op.dirname(__file__), '..', '..', 'fiff', 'tests',
+                        'data', 'test_ctf_comp_raw.fif')
+
 
 test_info = {'ch_names': ['ICA 001', 'ICA 002', 'EOG 061'],
  'chs': [{'cal': 1,
@@ -70,6 +89,8 @@ def test_io_layout_lout():
     assert_array_almost_equal(layout.pos, layout_read.pos, decimal=2)
     assert_true(layout.names, layout_read.names)
 
+    print layout  # test repr
+
 
 def test_io_layout_lay():
     """Test IO with .lay files"""
@@ -105,3 +126,81 @@ def test_make_grid_layout():
     assert_array_equal(lout_new.kind, tmp_name)
     assert_array_equal(lout_orig.pos, lout_new.pos)
     assert_array_equal(lout_orig.names, lout_new.names)
+
+
+def test_find_layout():
+    """Test finding layout"""
+    with warnings.catch_warnings(True) as w:
+        find_layout(chs=test_info['chs'])
+        assert_true(w[0].category == DeprecationWarning)
+    with warnings.catch_warnings(True) as w:
+        find_layout(test_info['chs'])
+        assert_true(w[0].category == DeprecationWarning)
+    assert_raises(ValueError, find_layout, dict())
+    assert_raises(ValueError, find_layout, test_info, ch_type='meep')
+        
+    sample_info = Raw(fif_fname).info
+    grads = pick_types(sample_info, meg='grad')
+    sample_info2 = pick_info(sample_info, grads)
+    
+    mags = pick_types(sample_info, meg='mag')
+    sample_info3 = pick_info(sample_info, mags)
+    
+    # mock new convention
+    sample_info4 = copy.deepcopy(sample_info)
+    for ii, name in enumerate(sample_info4['ch_names']):
+        new = name.replace(' ', '')
+        sample_info4['ch_names'][ii] = new
+        sample_info4['chs'][ii]['ch_name'] = new 
+
+    mags = pick_types(sample_info, meg=False, eeg=True)
+    sample_info5 = pick_info(sample_info, mags)
+
+    lout = find_layout(sample_info, ch_type=None)
+    assert_true(lout.kind == 'Vectorview-all')
+    assert_true(all(' ' in k for k in lout.names))
+
+    lout = find_layout(sample_info2, ch_type='meg')
+    assert_true(lout.kind == 'Vectorview-all')
+    
+    # test new vector-view
+    lout = find_layout(sample_info4, ch_type=None)
+    assert_true(lout.kind == 'Vectorview-all')
+    assert_true(all(not ' ' in k for k in lout.names))
+    
+    lout = find_layout(sample_info, ch_type='grad')
+    assert_true(lout.kind == 'Vectorview-grad')
+    lout = find_layout(sample_info2)
+    assert_true(lout.kind == 'Vectorview-grad')
+    lout = find_layout(sample_info2, ch_type='grad')
+    assert_true(lout.kind == 'Vectorview-grad')
+    lout = find_layout(sample_info2, ch_type='meg')
+    assert_true(lout.kind == 'Vectorview-all')
+
+    
+    lout = find_layout(sample_info, ch_type='mag')
+    assert_true(lout.kind == 'Vectorview-mag')
+    lout = find_layout(sample_info3)
+    assert_true(lout.kind == 'Vectorview-mag')
+    lout = find_layout(sample_info3, ch_type='mag')
+    assert_true(lout.kind == 'Vectorview-mag')
+    lout = find_layout(sample_info3, ch_type='meg')
+    assert_true(lout.kind == 'Vectorview-all')
+    # 
+    lout = find_layout(sample_info, ch_type='eeg')
+    assert_true(lout.kind == 'EEG')
+    lout = find_layout(sample_info5)
+    assert_true(lout.kind == 'EEG')
+    lout = find_layout(sample_info5, ch_type='eeg')
+    assert_true(lout.kind == 'EEG')
+    # no common layout, 'meg' option not supported
+
+    fname_bti_raw = op.join(bti_dir, 'exported4D_linux.fif')
+    lout = find_layout(Raw(fname_bti_raw).info)
+    assert_true(lout.kind == 'magnesWH3600')
+    
+    lout = find_layout(Raw(fname_ctf_raw).info)
+    assert_true(lout.kind == 'CTF-275')
+    
+    sample_info5['dig'] = []
+    assert_raises(RuntimeError, find_layout, sample_info5)